% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{hillclimbplot}
\alias{hillclimbplot}
\title{Plot the hillclimber trajectory}
\usage{
hillclimbplot(x, ...)
}
\arguments{
\item{x}{An object of class 'slouch'}

\item{...}{Additional arguments passed to 'plot.default(...)'}
}
\description{
Plot the hillclimber trajectory
}
\examples{
library(slouch)
library(ape)

data(neocortex)
data(artiodactyla)

neocortex <- neocortex[match(artiodactyla$tip.label, neocortex$species), ]

m0 <- slouch.fit(phy = artiodactyla,
                species = neocortex$species,
                response = neocortex$neocortex_area_mm2_log_mean,
                me.response = neocortex$neocortex_se_squared,
                hillclimb = TRUE)
                
hillclimbplot(m0)
}
