% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{hillclimbplot.slouch}
\alias{hillclimbplot.slouch}
\title{Plot the hillclimber trajectory}
\usage{
\method{hillclimbplot}{slouch}(x, ...)
}
\arguments{
\item{x}{An object of class 'slouch'}

\item{...}{Additional arguments passed to 'plot.default(...)'}
}
\description{
Plot the hillclimber trajectory
}
