% --- Source file: slouch-internal.Rd ---
\name{slouch-internal}
\alias{arrange.tree}
\alias{BiasCor}
\alias{branch.times}
\alias{DesMat}
\alias{DesMeCov}
\alias{DesPhyCov}
\alias{distance.matrix}
\alias{epochs}
\alias{ExpectedDesign}
\alias{fast.svd}
\alias{make.states}
\alias{n.ambig}
\alias{no.me.sigma.X.estimate}
\alias{nsmall.svd}
\alias{parse.tree}
\alias{pedigree}
\alias{positive.svd}
\alias{pseudoinverse}
\alias{psmall.svd}
\alias{regimes}
\alias{ResidV}
\alias{set.of.regimes}
\alias{terminal.twigs}
\alias{tree.layout}
\alias{tsia}
\alias{tsja}
\alias{weight.matrix}
\title{Internal Slouch Functions}
\description{
  Internal Slouch functions.
}
\note{
  These are not generally called by the user (or in some cases are just
  waiting for proper documentation to be written). They are used by other   slouch functions and in many cases have been borrowed and modified form other packages, particularly the ouch package. 
}
\keyword{internal}
