% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{lseq}
\alias{lseq}
\title{Logarithmically spaced sequence}
\usage{
lseq(from = 1, to = 1e+05, length.out = 6)
}
\arguments{
\item{from}{the starting value of the sequence. Must be positive.}

\item{to}{the terminal value of the sequence. Must be larger than input to "from".}

\item{length.out}{desired length of the sequence. Must not be negative.}
}
\value{
A sequence of logarithmically spaced numbers.
}
\description{
Logarithmically spaced sequence
}
\examples{
lseq(1, 1000, length.out = 4)
}
