% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slouch.R
\name{model.fit.dev2}
\alias{model.fit.dev2}
\title{Title}
\usage{
model.fit.dev2(phy, species = NULL, half_life_values = NULL,
  vy_values = NULL, response, me.response = NULL, fixed.fact = NULL,
  fixed.cov = NULL, me.fixed.cov = NULL, mecov.fixed.cov = NULL,
  random.cov = NULL, me.random.cov = NULL, mecov.random.cov = NULL,
  estimate.Ya = FALSE, estimate.bXa = FALSE, support = 2,
  convergence = 1e-06, nCores = 1, hillclimb = FALSE,
  hillclimb_start = NULL, verbose = FALSE)
}
\arguments{
\item{phy}{an object of class 'phylo', must be rooted.}

\item{species}{.}

\item{half_life_values}{A vector of candidate phylogenetic half-life values to be evaluated in grid search.}

\item{vy_values}{A vector of candidate stationary variances for the response trait, to be evaluated in grid search.}

\item{response}{A numeric vector of a trait to be treated as response variable}

\item{me.response}{Numeric vector of the observational variances of each response trait. E.g if response is a mean trait value, me.response is the within-species SE of the mean.}

\item{fixed.fact}{Factor}

\item{fixed.cov}{Direct effect independent variables}

\item{me.fixed.cov}{Observational variances for direct effect independent variables}

\item{mecov.fixed.cov}{.}

\item{random.cov}{Independent variables each modeled as a brownian motion}

\item{me.random.cov}{Observational variances for the brownian covariates}

\item{mecov.random.cov}{.}

\item{estimate.Ya}{Boolean. If true, the intercept K = 1 is expanded to Ya = exp(-a*T.term) and b0 = 1-exp(-a*T.term). If models with categorical covariates are used, this will instead estimate a separate primary optimum for the root niche, "Ya". This only makes sense for non-ultrametric trees. If the tree is ultrametric, the model matrix becomes singular.}

\item{estimate.bXa}{Boolean. If true, bXa = bXa = 1-exp(-a*T.term) - (1-(1-exp(-a*T.term))/(a*T.term)) is added to the model matrix, estimating b*Xa. Same requirements as for estimating Ya.}

\item{support}{A scalar indicating the size of the support set, defaults to 2 units of log-likelihood.}

\item{convergence}{Threshold of iterative GLS estimation, when beta is considered converged.}

\item{nCores}{Use multiple CPU cores in grid-search. If 2 or more cores are used, all print statements are silenced during grid search. Optional, defaults to 1.}

\item{hillclimb}{TRUE/FALSE whether to use hillclimb parameter estimation routine.}

\item{hillclimb_start}{Numeric vector of length 2, c(hl, vy), to specify where the hillclimber routine starts. Optional.}

\item{verbose}{if TRUE, prints each iteration of parameter search. Default FALSE.}
}
\value{
An object of class 'slouch'
}
\description{
Title
}

