% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.slouch}
\alias{plot.slouch}
\title{Plot slouch-objects}
\usage{
plot.slouch(x, ...)
}
\arguments{
\item{x}{An object of class 'slouch'}

\item{...}{Additional parameters passed to persp()}
}
\description{
Graphical plot of parameter space traversed in order to find ML-estimate of the model.
}
