% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.slouch}
\alias{print.slouch}
\title{Print slouch-objects}
\usage{
\method{print}{slouch}(x, ...)
}
\arguments{
\item{x}{An object of class 'slouch'}

\item{...}{Additional arguments, unused.}
}
\description{
Print slouch-objects
}
