% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{regimeplot.slouch}
\alias{regimeplot.slouch}
\title{Plot the internal regimes for a given fitted model}
\usage{
\method{regimeplot}{slouch}(x, ...)
}
\arguments{
\item{x}{an object of class 'slouch'}

\item{...}{additional parameters passed to plot.phylo(...)}
}
\value{
nothing
}
\description{
Plot the internal regimes for a given fitted model
}
