% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slouchbtk.R
\docType{package}
\name{slouchbtk-package}
\alias{slouchbtk}
\alias{slouchbtk-package}
\title{SLOUCH: Stochastic Linear Ornstein Uhlenbeck Comparative Hypotheses}
\description{
More about what it does (maybe more than one line)
}
\details{
Forked by bjørn
}

