% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slouch1.R
\docType{package}
\name{slouchexp-package}
\alias{slouchexp}
\alias{slouchexp-package}
\title{SLOUCH: Stochastic Linear Ornstein Uhlenbeck Comparative Hypotheses}
\description{
SLOUCH: Stochastic Linear Ornstein Uhlenbeck Comparative Hypotheses
}

