% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artiodactyla}
\alias{artiodactyla}
\title{Artiodactyla Phylogeny}
\format{
An object of class \code{"phylo"}, from package \code{"ape"}.
}
\usage{
data(artiodactyla)
}
\description{
A phylogenetic tree of Artiodactyla (even-toed ungulates).
}
\examples{
data(artiodactyla)
library(ape)
plot(artiodactyla)

# Note: This tree also has regime information for each internal node in the tree:
print(artiodactyla$node.label)

}
\references{
The tree is taken from Toljagic et al. (2017).

\itemize{
\item{Toljagić, O., Voje, K. L., Matschiner, M., Liow, L. H., & Hansen, T. F. (2017). Millions of years behind: Slow adaptation of ruminants to grasslands. Systematic Biology, (318). https://doi.org/10.1093/sysbio/syx059}
}
}
\keyword{datasets}
