% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{fitch}
\alias{fitch}
\title{Title}
\usage{
fitch(tree.data, niche, deltran = FALSE, acctran = FALSE, root = NULL)
}
\arguments{
\item{tree.data}{A data.frame including columns "ancestor", "nodes".}

\item{niche}{Factor variable to be used in ancestral reconstruction.}

\item{deltran}{.}

\item{acctran}{.}

\item{root}{A priori specification of root niche.}
}
\value{
data.frame
}
\description{
Title
}

