% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{fitch}
\alias{fitch}
\title{Title}
\usage{
fitch(tree.data, niche, deltran = FALSE, acctran = FALSE, root = NULL)
}
\arguments{
\item{tree.data}{}

\item{niche}{}

\item{deltran}{}

\item{acctran}{}

\item{root}{}
}
\description{
Title
}

