% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.slouch}
\alias{logLik.slouch}
\title{Title}
\usage{
\method{logLik}{slouch}(object, ...)
}
\arguments{
\item{object}{An object of class 'slouch'}

\item{...}{Additional arguments.}
}
\value{
An object of class 'logLik'
}
\description{
Title
}
