% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slouch1.R
\name{model.fit.dev2}
\alias{model.fit.dev2}
\title{Title}
\usage{
model.fit.dev2(topology, times, half_life_values, vy_values, response,
  me.response = NULL, fixed.fact = NULL, fixed.cov = NULL,
  me.fixed.cov = NULL, mecov.fixed.cov = NULL, random.cov = NULL,
  me.random.cov = NULL, mecov.random.cov = NULL, ultrametric = TRUE,
  intercept = if (ultrametric == TRUE) "root" else NULL, support = 2,
  convergence = 1e-06, plot.angle = 30, parallel.compute = FALSE,
  hillclimb = FALSE)
}
\arguments{
\item{plot.angle}{}
}
\description{
Title
}

