% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLOCH_BTK_09.07.2016.R
\name{model.fit}
\alias{model.fit}
\title{Title}
\usage{
model.fit(topology, times, half_life_values, vy_values, response,
  me.response = NULL, fixed.fact = NULL, fixed.cov = NULL,
  me.fixed.cov = NULL, mecov.fixed.cov = NULL, random.cov = NULL,
  me.random.cov = NULL, mecov.random.cov = NULL, intercept = "root",
  ultrametric = TRUE, support = NULL, convergence = NULL,
  plot.angle = 30)
}
\arguments{
\item{plot.angle}{}
}
\description{
Title
}

