% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{np.regression}
\alias{np.regression}
\title{Title}
\usage{
np.regression(response, me.response, predictor, me.predictor,
  convergence = NULL)
}
\arguments{
\item{response}{}

\item{me.response}{}

\item{predictor}{}

\item{me.predictor}{}

\item{convergence}{}
}
\description{
Title
}

