% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{ouch2slouch}
\alias{ouch2slouch}
\title{Title}
\usage{
ouch2slouch(tree)
}
\arguments{
\item{tree}{An object of class 'ouchtree'}
}
\value{
A 'slouchtree', a data.frame with columns "nodes", "ancestor", "times" and "species".
}
\description{
Title
}

