% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.slouch}
\alias{print.slouch}
\title{Print Model Summary}
\usage{
\method{print}{slouch}(x, ...)
}
\arguments{
\item{x}{An object of class 'slouch'}

\item{...}{Additional arguments, unused.}
}
\description{
Print Model Summary
}
\examples{
data(artiodactyla)
data(neocortex)

neocortex <- neocortex[match(artiodactyla$tip.label, neocortex$species), ]

m0 <- slouch.fit(phy = artiodactyla,
                 hl_values = seq(0.001, 50, length.out = 15),
                 vy_values = seq(0.001, 3, length.out = 15),
                 species = neocortex$species,
                 response = neocortex$body_mass_g_log_mean,
                 mv.response = neocortex$body_mass_g_log_varmean,
                 fixed.fact = neocortex$diet)
                 
print(m0) 
}
