% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slouch.R
\docType{package}
\name{slouch-package}
\alias{slouch}
\alias{slouch-package}
\title{SLOUCH: Stochastic Linear Ornstein Uhlenbeck Comparative Hypotheses}
\description{
SLOUCH: Stochastic Linear Ornstein Uhlenbeck Comparative Hypotheses
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/kopperud/slouch}
  \item Report bugs at \url{http://github.com/kopperud/slouch/issues}
}

}
