% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{slouch.fit}
\alias{slouch.fit}
\title{Function to fit Ornstein-Uhlenbeck models}
\usage{
slouch.fit(phy, species = NULL, hl_values = NULL, vy_values = NULL,
  response, me.response = NULL, fixed.fact = NULL, fixed.cov = NULL,
  me.fixed.cov = NULL, mecov.fixed.cov = NULL, random.cov = NULL,
  me.random.cov = NULL, mecov.random.cov = NULL, estimate.Ya = FALSE,
  estimate.bXa = FALSE, support = 2, convergence = 1e-06, nCores = 1,
  hillclimb = FALSE, hillclimb_start = NULL, lower = c(0, 0),
  upper = c(Inf, Inf), verbose = FALSE)
}
\arguments{
\item{phy}{an object of class 'phylo', must be rooted.}

\item{species}{a character vector of species tip labels, typically the "species" column in a data frame. This column needs to be an exact match and same order as phy$tip.label}

\item{hl_values}{a vector of candidate phylogenetic half-life values to be evaluated in grid search. Optional.}

\item{vy_values}{a vector of candidate stationary variances for the response trait, to be evaluated in grid search. Optional.}

\item{response}{a numeric vector of a trait to be treated as response variable}

\item{me.response}{numeric vector of the observational variances of each response trait. E.g if response is a mean trait value, me.response is the within-species squared standard error of the mean.}

\item{fixed.fact}{factor of regimes on the terminal edges of the tree, in same order as species. If this is used, phy$node.label needs to be filled with the corresponding internal node regimes, in the order of node indices (root: n+1),(n+2),(n+3), ...}

\item{fixed.cov}{Direct effect independent variables}

\item{me.fixed.cov}{Observational variances for direct effect independent variables. Must be the same shape as fixed.cov}

\item{mecov.fixed.cov}{.}

\item{random.cov}{Independent variables each modeled as a brownian motion}

\item{me.random.cov}{Observational variances for the brownian covariates. Must be the same shape as random.cov}

\item{mecov.random.cov}{.}

\item{estimate.Ya}{a logical value indicathing whether "Ya" should be estimated. If true, the intercept K = 1 is expanded to Ya = exp(-a*t) and b0 = 1-exp(-a*t). If models with categorical covariates are used, this will instead estimate a separate primary optimum for the root niche, "Ya". This only makes sense for non-ultrametric trees. If the tree is ultrametric, the model matrix becomes singular.}

\item{estimate.bXa}{a logical value indicathing whether "bXa" should be estimated. If true, bXa = 1-exp(-a*t) - (1-(1-exp(-a*t))/(a*t)) is added to the model matrix, estimating b*Xa. Same requirements as for estimating Ya.}

\item{support}{a scalar indicating the size of the support set, defaults to 2 units of log-likelihood.}

\item{convergence}{threshold of iterative GLS estimation for when beta is considered to be converged.}

\item{nCores}{number of CPU cores used in grid-search. If 2 or more cores are used, all print statements are silenced during grid search. If performance is critical it is recommended to compile and link R to a multithreaded BLAS, since most of the heavy computations are common matrix operations. Even if a singlethreaded BLAS is used, this may or may not improve performance, and performance may vary with OS.}

\item{hillclimb}{logical, whether to use hillclimb parameter estimation routine or not. This routine (L-BFGS-B from optim()) may be combined with the grid-search, in which case it will on default start on the sigma and halflife for the local ML found by the grid-search.}

\item{hillclimb_start}{numeric vector of length 2, c(hl, vy), to specify where the hillclimber routine starts.}

\item{lower}{lower bounds for the optimization routine, defaults to c(0,0). First entry in vector is half-life, second is stationary variance. When running direct effect models without observational error, it may be useful to specify a positive lower bounds for the stationary variance, e.g c(0, 0.001), since the residual variance-covariance matrix is degenerate when sigma = 0.}

\item{upper}{upper bounds for the optimization routine, defaults to c(Inf, Inf).}

\item{verbose}{a logical value indicating whether to print a summary in each iteration of parameter search. May be useful when diagnosing unexpected behaviour or crashes.}
}
\value{
An object of class 'slouch', essentially a list with the following fields:

\item{parameter_space}{a list of the entire parameter space traversed by the grid search and the hillclimber as applicable}
\item{tree}{a list including the following:
\itemize{
\item{phy - an object of class 'phy'}
\item{T.term - a numeric vector including the time from the root of the tree to the tip, for all taxa 1,2,3... n}

}
}
\item{Lorem ipsum dolorem}{Yes \itemize{\item hello \item there}}
}
\description{
Function to fit Ornstein-Uhlenbeck models
}
