% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{slouchtree.plot}
\alias{slouchtree.plot}
\title{Title}
\usage{
slouchtree.plot(topology, times, names = NULL, regimes = NULL, cex = NULL,
  lwd = NULL, reg.col = NULL)
}
\arguments{
\item{reg.col}{}
}
\description{
Title
}

