% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{slouchtree.plot}
\alias{slouchtree.plot}
\title{Title}
\usage{
slouchtree.plot(topology, times, names = NULL, regimes = NULL, cex = NULL,
  lwd = NULL, reg.col = NULL)
}
\arguments{
\item{topology}{Ancestor vector}

\item{times}{Time from root to node}

\item{names}{A vector of names for each node to be labelled. Optional}

\item{regimes}{A vector of regimes to be painted on the tree. Optional}

\item{cex}{Font size}

\item{lwd}{Line width}

\item{reg.col}{Colour}
}
\description{
Title
}

